// test/integration/test/ports/engines/queries/filters/ported_filters/_fixture/generated/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgAdapterQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly ModelA: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly idTest: CodecTypes['pg/text@1']['output'] | null;
      readonly optString: CodecTypes['pg/text@1']['output'] | null;
      readonly optInt: CodecTypes['pg/int4@1']['output'] | null;
      readonly optFloat: CodecTypes['pg/float8@1']['output'] | null;
      readonly optBoolean: CodecTypes['pg/bool@1']['output'] | null;
      readonly optDateTime: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly optEnum: 'A' | 'B' | null;
      readonly b_id: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly ModelB: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly ModelA: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly idTest: CodecTypes['pg/text@1']['input'] | null;
      readonly optString: CodecTypes['pg/text@1']['input'] | null;
      readonly optInt: CodecTypes['pg/int4@1']['input'] | null;
      readonly optFloat: CodecTypes['pg/float8@1']['input'] | null;
      readonly optBoolean: CodecTypes['pg/bool@1']['input'] | null;
      readonly optDateTime: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly optEnum: 'A' | 'B' | null;
      readonly b_id: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly ModelB: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly modelA: {
      readonly b_id: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly idTest: CodecTypes['pg/text@1']['output'] | null;
      readonly optBoolean: CodecTypes['pg/bool@1']['output'] | null;
      readonly optDateTime: CodecTypes['pg/timestamptz@1']['output'] | null;
      readonly optEnum: 'A' | 'B' | null;
      readonly optFloat: CodecTypes['pg/float8@1']['output'] | null;
      readonly optInt: CodecTypes['pg/int4@1']['output'] | null;
      readonly optString: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly modelB: {
      readonly id: CodecTypes['pg/text@1']['output'];
      readonly int: CodecTypes['pg/int4@1']['output'] | null;
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly modelA: {
      readonly b_id: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly idTest: CodecTypes['pg/text@1']['input'] | null;
      readonly optBoolean: CodecTypes['pg/bool@1']['input'] | null;
      readonly optDateTime: CodecTypes['pg/timestamptz@1']['input'] | null;
      readonly optEnum: 'A' | 'B' | null;
      readonly optFloat: CodecTypes['pg/float8@1']['input'] | null;
      readonly optInt: CodecTypes['pg/int4@1']['input'] | null;
      readonly optString: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly modelB: {
      readonly id: CodecTypes['pg/text@1']['input'];
      readonly int: CodecTypes['pg/int4@1']['input'] | null;
    };
  };
};
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;